// src/models/Rmc/Rmc.ts
export interface Rmc {
    id: string
    tenant_id: string

    name_client: string;
    calculation_base: string;
    committed_value: string;
    bank_name: string;
    cpf_client: string;
    address_client: string;
    city_client: string;
    inclusion_date: string;
    contract_value: string;
    installment_value: string;
    contract_number: string;
    type_process: string;
}